Call posix_fadvise less frequently
This improves a bit total time spent in system call by having a threshold
when posix_fadvise() is called. It is called only when threshold amount of
data is read. Threshold size 8 MB was selected because it performed a little
better than 2 MB and 16 MB did not improve much. After all 8 MB is modest
amount of buffer or page cache pollution when typically reading gigabytes of
data.

Signed-off-by: Jarkko Nikula <jarkko.nikula@bitmer.com>